Move builders out of Util namespace
