+++ src/KubeOps/Operator/Builder/OperatorBuilder.cs
@@ -7,10 +7,12 @@ using k8s.Models;
 using KubeOps.Operator.Caching;
 using KubeOps.Operator.Controller;
 using KubeOps.Operator.DevOps;
+using KubeOps.Operator.Entities;
 using KubeOps.Operator.Events;
 using KubeOps.Operator.Finalizer;
 using KubeOps.Operator.Kubernetes;
 using KubeOps.Operator.Leadership;
+using KubeOps.Operator.Rbac;
 using KubeOps.Operator.Serialization;
 using KubeOps.Operator.Services;
 using KubeOps.Operator.Util;
