+++ src/KubeOps/Operator/Commands/Generators/CrdGenerator.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using k8s.Models;
 using KubeOps.Operator.Commands.CommandHelpers;
+using KubeOps.Operator.Entities;
 using KubeOps.Operator.Entities.Kustomize;
 using KubeOps.Operator.Serialization;
 using KubeOps.Operator.Util;
