+++ tests/KubeOps.Test/Operator/Util/CrdGenerator.Test.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
 using k8s.Models;
+using KubeOps.Operator.Entities;
 using KubeOps.Operator.Services;
 using KubeOps.Operator.Util;
 using KubeOps.Test.TestEntities;
